remove more unused stuff
