+++ packages/client-core/i18n/en/admin.json
@@ -232,7 +232,9 @@
       "entity": "Entity",
       "authStrategies": "Authentication Strategies",
       "userName": "User Name",
+      "local": "Local",
       "oauth": "OAuth",
+      "password": "Password",
       "defaults": "Defaults",
       "host": "Host",
       "protocol": "Protocol",
