+++ packages/client-core/i18n/en/social.json
@@ -114,7 +114,6 @@
     "login": "Login",
     "notHavingAccount": "Don't have an account?",
     "signUp": "Sign Up",
-    "forgotPassword": "Forgot password?",
     "back": "Back",
     "or": "OR"
   },
