--- packages/client-core/src/components/World/loading.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Loading = () => (
-  <div>
-    <h3>Loading...</h3>
-  </div>
-)
-
-export default Loading
