+++ packages/client-core/src/user/components/Auth/MagicLinkEmail.tsx
