+++ packages/client/src/pages/auth/authRoutes.tsx
@@ -5,7 +5,6 @@ import { Route, Routes } from 'react-router-dom'
 import { LoadingCircle } from '@etherealengine/client-core/src/components/LoadingCircle'
 
 import $confirm from './confirm'
-import $forgotpassword from './forgotpassword'
 import $magiclink from './magiclink'
 import $discord from './oauth/discord'
 import $facebook from './oauth/facebook'
@@ -25,7 +24,6 @@ const AuthRoutes = () => {
         <Route path="oauth/linkedin" element={<$linkedin />} />
         <Route path="oauth/twitter" element={<$twitter />} />
         <Route path="confirm" element={<$confirm />} />
-        <Route path="forgotpassword" element={<$forgotpassword />} />
         <Route path="magiclink" element={<$magiclink />} />
       </Routes>
     </Suspense>
