--- packages/client/src/pages/auth/forgotpassword.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-import ForgotPassword from '@etherealengine/client-core/src/user/components/Auth/ForgotPassword'
-
-export const ForgotPasswordPage = () => <ForgotPassword />
-
-export default ForgotPasswordPage
